src/main/java/ru/job4j/interaction/SimpleBlockingQueue.java
package ru.job4j.interaction;

import net.jcip.annotations.GuardedBy;

import java.util.LinkedList;
import java.util.Queue;

public class SimpleBlockingQueue<T> {
    @GuardedBy("this")
    private Queue<T> queue = new LinkedList<>();

    synchronized public void offer(T value) {
        while (!queue.offer(value)) {
            try {
                this.wait();
            } catch (InterruptedException e) {
                Thread.currentThread().interrupt();
            }
        }
        this.notifyAll();
    }

    synchronized public T poll() {
        T item;
        while ((item = queue.poll()) == null) {
            try {
                this.wait();
            } catch (InterruptedException e) {
                Thread.currentThread().interrupt();
            }
        }
        this.notifyAll();
        return item;
    }
}
